Added command runner, moved some tests to specific files
