+++ .gitignore
@@ -87,6 +87,7 @@ ENV/
 
 # Rope project settings
 .ropeproject
+.idea
 
 # Build systems
 node_modules/
