+++ devel.dockerfile
@@ -9,5 +9,5 @@ RUN pip install -U pip
 RUN pip install -I -e .[develop] --process-dependency-links
 
 WORKDIR /run/service/src/apof
-ENTRYPOINT ["python", "manage.py"]
-CMD ["runserver", "0.0.0.0:8000"]
+ENTRYPOINT ["apofcmd"]
+CMD ["django","runserver","0.0.0.0:8000"]
