+++ docker-compose.yml
@@ -12,6 +12,8 @@ services:
       - "8000:8000"
     depends_on:
       - db
+    environment:
+      - DJANGO_SETTINGS_MODULE=apof.service.settings.development
   db:
     image: mysql:5.6
     environment:
