+++ src/apof/baskets/tests.py
@@ -2,8 +2,6 @@ from decimal import Decimal
 from mock import MagicMock, patch
 
 from django.apps.config import AppConfig
-from django.contrib.admin.filters import SimpleListFilter, ListFilter
-from django.contrib.admin.options import ModelAdmin, BaseModelAdmin
 from django.contrib.auth.models import Permission, User
 from django.core.files import File
 from django.shortcuts import reverse
@@ -13,6 +11,7 @@ from baskets.apps import BasketsConfig
 from baskets.admin import RestaurantListFilter, OrderAdmin, BasketAdmin
 from baskets.models import Basket, Order
 from baskets.templatetags.get_item import get_item
+from baskets.tests.utils import OrderTestMixin
 from baskets.views import OrderDeleteView, OrderListView
 from menus.models import (
     Ingredient,
@@ -33,43 +32,6 @@ class BasketTestCase(TestCase):
         self.assertEqual(str(basket), 'Basket christopher')
 
 
-class OrderTestMixin:
-
-    @patch('django.core.files.storage.default_storage._wrapped')
-    def setUp(self, mocked_storage):
-        file_mock = MagicMock(spec=File, name='FileMock')
-        file_mock.name = 'test1.jpg'
-        restaurant = Restaurant.objects.create(name='test restaurant2', logo=file_mock)
-        menu = Menu.objects.create(restaurant=restaurant)
-        size = Size.objects.create(menu=menu, value=15, value_unit='cm')
-        test_meal = Meal.objects.create(menu=menu, name='test_meal')
-        test1_topping = Topping.objects.create(
-            menu=menu,
-            ingredient=Ingredient.objects.create(
-                name='test1_ingredient'
-            )
-        )
-        test2_topping = Topping.objects.create(
-            menu=menu,
-            ingredient=Ingredient.objects.create(
-                name='test2_ingredient'
-            )
-        )
-        Price.objects.bulk_create([
-            Price(value=22.99, size=size, content_object=test_meal),
-            Price(value=2.5, size=size, content_object=test1_topping),
-            Price(value=1.25, size=size, content_object=test2_topping)
-        ])
-        self.order = Order.objects.create(
-            basket=Basket.objects.create(owner=User.objects.get(username='christopher')),
-            meal=test_meal,
-            size=size
-        )
-        self.order.toppings.set([test1_topping, test2_topping])
-        self.data_mock = MagicMock()
-        self.test_order = OrderAdmin(self.data_mock, self.data_mock)
-
-
 class OrderTestCase(OrderTestMixin, TestCase):
     fixtures = ['test_user_data.json']
 
@@ -87,61 +49,6 @@ class OrderTestCase(OrderTestMixin, TestCase):
     def test_get_total_price(self):
         expected_total_price = Decimal('26.74')
         self.assertEqual(self.order.get_total_price(), expected_total_price)
-
-
-class OrderListViewTestCase(OrderTestMixin, TestCase):
-    fixtures = ['test_user_data.json']
-
-    def test_constants(self):
-        self.assertEqual(OrderListView.model, Order)
-        self.assertEqual(OrderListView.permission_required, ('baskets.delete_order', ))
-        self.assertTrue(OrderListView.raise_exception)
-        self.assertEqual(OrderListView.ordering, ('meal__menu__restaurant', ))
-
-    def test_anonymous_user_gets_403(self):
-        response = self.client.get(reverse('order-list'))
-        self.assertEqual(response.status_code, 403)
-
-    def test_logged_user_without_permission_gets_403(self):
-        self.client.force_login(User.objects.get(username='christopher'))
-
-        response = self.client.get(reverse('order-list'))
-        self.assertEqual(response.status_code, 403)
-
-    def test_logged_user_with_permission(self):
-        user = User.objects.get(username='christopher')
-        user.user_permissions.add(Permission.objects.get(codename='delete_order'))
-        self.client.force_login(user)
-
-        response = self.client.get(reverse('order-list'))
-        self.assertEqual(response.status_code, 200)
-
-
-class OrderDeleteViewTestCase(OrderTestMixin, TestCase):
-    fixtures = ['test_user_data.json']
-
-    def test_constants(self):
-        self.assertEqual(OrderDeleteView.model, Order)
-        self.assertEqual(OrderDeleteView.permission_required, ('baskets.delete_order', ))
-        self.assertTrue(OrderDeleteView.raise_exception)
-
-    def test_anonymous_user_gets_403(self):
-        response = self.client.get(reverse('order-delete', kwargs={'pk': self.order.pk}))
-        self.assertEqual(response.status_code, 403)
-
-    def test_logged_user_without_permission_gets_403(self):
-        self.client.force_login(User.objects.get(username='christopher'))
-
-        response = self.client.get(reverse('order-delete', kwargs={'pk': self.order.pk}))
-        self.assertEqual(response.status_code, 403)
-
-    def test_logger_user_with_permission(self):
-        user = User.objects.get(username='christopher')
-        user.user_permissions.add(Permission.objects.get(codename='delete_order'))
-        self.client.force_login(user)
-
-        response = self.client.get(reverse('order-delete', kwargs={'pk': self.order.pk}))
-        self.assertEqual(response.status_code, 200)
 
 
 class GetItemTagTestCase(TestCase):
@@ -215,19 +122,8 @@ class AddToBasketViewTestCase(TestCase):
 
 
 class AdminTestCase(OrderTestMixin, TestCase):
-    fixtures = ['test_user_data.json']
 
-    def test_admin_restaurant_list(self):
-        testFilter = RestaurantListFilter(
-            self.data_mock,
-            self.data_mock,
-            self.data_mock,
-            self.data_mock
-        )
-        data1 = testFilter.lookups(self.data_mock, self.data_mock)
-        data2 = testFilter.queryset(self.data_mock, data1)
-        self.assertEqual(data1[0][1], 'test restaurant2')
-        self.assertEqual(data2[0][1], 'test restaurant2')
+    fixtures = ['test_user_data.json']
 
     def test_admin_get_meal_name(self):
         data = self.test_order.get_meal_name(self.order)
@@ -244,21 +140,6 @@ class AdminTestCase(OrderTestMixin, TestCase):
     def test_admin_get_toppings(self):
         data = self.test_order.get_toppings(self.order)
         self.assertEqual(data, 'test1_ingredient, test2_ingredient')
-
-    def test_OrderAdmin_mro(self):
-        orderMRO = OrderAdmin.__mro__
-        expectedMRO = (OrderAdmin, ModelAdmin, BaseModelAdmin, object)
-        self.assertEqual(orderMRO, expectedMRO)
-
-    def test_BasketAdmin_mro(self):
-        basketMRO = BasketAdmin.__mro__
-        expectedMRO = (BasketAdmin, ModelAdmin, BaseModelAdmin, object)
-        self.assertEqual(basketMRO, expectedMRO)
-
-    def test_RestaurantListFilter_mro(self):
-        filterMRO = RestaurantListFilter.__mro__
-        expectedMRO = (RestaurantListFilter, SimpleListFilter, ListFilter, object)
-        self.assertEqual(filterMRO, expectedMRO)
 
 
 class AppsTestCase(TestCase):
