+++ src/apof/menus/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from menus import views
+from apof.menus import views
 
 urlpatterns = [
     url(r'^(?P<restaurant_pk>[0-9]+)/$', views.MealListView.as_view(), name='meal-list')
