+++ src/apof/service/urls.py
@@ -22,10 +22,10 @@ from django.views.generic.base import RedirectView
 urlpatterns = [
     url(r'^$', RedirectView.as_view(url='/portal/')),
     url(r'^admin/', admin.site.urls),
-    url(r'^baskets/', include('baskets.urls')),
-    url(r'^menus/', include('menus.urls')),
-    url(r'^portal/', include('portal.urls')),
-    url(r'^restaurants/', include('restaurants.urls'))
+    url(r'^baskets/', include('apof.baskets.urls')),
+    url(r'^menus/', include('apof.menus.urls')),
+    url(r'^portal/', include('apof.portal.urls')),
+    url(r'^restaurants/', include('apof.restaurants.urls'))
 ]
 
 if settings.DEBUG:
